requirements.txt
minimalmodbus==2.1.1
pymodbus==3.11.4
pyserial==3.5
